+++ lib/rubocop/ast/builder.rb
@@ -45,6 +45,7 @@ module RuboCop
         resbody:      ResbodyNode,
         retry:        RetryNode,
         return:       ReturnNode,
+        self:         SelfNode,
         csend:        SendNode,
         send:         SendNode,
         str:          StrNode,