+++ spec/rubocop/cop/style/one_line_conditional_spec.rb
@@ -130,4 +130,15 @@ RSpec.describe RuboCop::Cop::Style::OneLineConditional do
       true ? break : 7
     RUBY
   end
+
+  it 'does not break when one of the branches contains a self keyword' do
+    expect_offense(<<~RUBY)
+      if true then self else 7 end
+      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Favor the ternary operator (`?:`) over `if/then/else/end` constructs.
+    RUBY
+
+    expect_correction(<<~RUBY)
+      true ? self : 7
+    RUBY
+  end
 end